requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.3.0
django-filter==23.3
drf-spectacular==0.26.5